+++ packages/typespec-client-generator-core/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./decorators.js";
-// export { $hasJSONConverter } from "./csharp.decorators.js";
 export * from "./interfaces.js";
 export * from "./lib.js";
 export * from "./public-utils.js";
